# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Visual Diagram Gen API",
    "version": "1.0.0",
    "description": "Backend API for the AI explainer animation tool. This document describes the public HTTP contract used by the frontend and MCP tooling."
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local development API" }
  ],
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "array", "items": {} }
            },
            "required": ["code", "message", "details"]
          }
        },
        "required": ["error"]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" }
        },
        "required": ["id", "name", "email"]
      },
      "AuthTokenResponse": {
        "type": "object",
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "token": { "type": "string" }
        },
        "required": ["user", "token"]
      },
...
      "Asset": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": {
            "type": "string",
            "enum": ["audio", "video", "image", "logo", "side_video"]
          },
          "originalName": { "type": "string" },
          "storageKey": { "type": "string" },
          "mimeType": { "type": "string" },
          "sizeBytes": { "type": "integer", "format": "int64" },
          "durationSeconds": { "type": ["number", "null"], "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "kind",
          "originalName",
          "storageKey",
          "mimeType",
          "sizeBytes",
          "createdAt"
        ]
      },
...
      "ItemsAssetResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Asset" }
          }
        },
        "required": ["items"]
      },
...
    },
    "parameters": {
...
      "AssetId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "FileTooLarge": {
        "description": "File too large",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "minLength": 2 },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 12 }
                },
                "required": ["name", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokenResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login and issue a JWT",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthTokenResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Request a password reset token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reset request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "message": { "type": "string" }
                  },
                  "required": ["ok", "message"]
                }
              }
            }
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "summary": "Reset a user password using a reset token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string", "minLength": 16 },
                  "password": { "type": "string", "minLength": 12 }
                },
                "required": ["token", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "message": { "type": "string" }
                  },
                  "required": ["ok", "message"]
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/auth/me": {
      "get": {
        "summary": "Get current authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "$ref": "#/components/schemas/User" }
                  },
                  "required": ["user"]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
    "/assets": {
      "post": {
        "summary": "Upload an asset",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "kind": {
                    "type": "string",
                    "enum": ["audio", "video", "image", "logo", "side_video"]
                  }
                },
                "required": ["file", "kind"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Asset created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Asset" }
              }
            }
          },
          "413": { "$ref": "#/components/responses/FileTooLarge" }
        }
      },
      "get": {
        "summary": "List assets",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "schema": {
              "type": "string",
              "enum": ["audio", "video", "image", "logo", "side_video"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ItemsAssetResponse" }
              }
            }
          }
        }
      }
    },
    "/assets/{id}/file": {
      "get": {
        "summary": "Download an asset file",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AssetId" }],
        "responses": {
          "200": {
            "description": "Binary asset stream",
            "content": { "audio/*": {}, "video/*": {}, "image/*": {} }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/assets/{id}": {
      "delete": {
        "summary": "Delete an asset",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AssetId" }],
        "responses": {
          "204": { "description": "Asset deleted" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  }
}